dashboard_generation/requirements.txt
black==22.3.0
jira==3.2.0
python-dotenv==0.20.0
